+++ README.md
@@ -69,6 +69,6 @@ If you are using SystemJS, you should also adjust your configuration to point to
 In your systemjs config file, `map` needs to tell the System loader where to look for `ng2-adsense`:
 ```js
 map: {
-  'ng2-adsense': 'node_modules/ng2-adsense/ngadsense.umd.js',
+  'ng2-adsense': 'node_modules/ng2-adsense/ng2-adsense.umd.js',
 }
 ```
