+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "ng2-adsense",
-  "version": "4.0.6",
+  "version": "4.0.8",
   "license": "MIT",
   "repository": "https://github.com/scttcper/ng2-adsense.git",
   "private": true,
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "node build.js",
+    "build": "rimraf dist && node build.js",
     "publish": "cd dist/packages-dist && npm publish",
     "test": "ng test",
     "ghpages": "ng build --prod --aot --build-optimizer --no-progress --baseHref='/ng2-adsense/'",
@@ -24,7 +24,7 @@
     "bootstrap": "^4.0.0-beta",
     "core-js": "^2.5.0",
     "rxjs": "^5.4.3",
-    "zone.js": "^0.8.16"
+    "zone.js": "^0.8.17"
   },
   "devDependencies": {
     "@angular/cli": "1.3.1",
@@ -32,7 +32,7 @@
     "@types/jasmine": "^2.5.54",
     "@types/node": "^8.0.24",
     "codelyzer": "^3.1.2",
-    "gzip-size": "^3.0.0",
+    "fs-extra": "^4.0.1",
     "jasmine-core": "^2.7.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^1.7.0",
@@ -41,9 +41,10 @@
     "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "^1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "pretty-size": "^2.0.0",
     "protractor": "^5.1.2",
     "rollup": "^0.48.2",
+    "rollup-plugin-cleanup": "^1.0.1",
+    "rollup-plugin-filesize": "^1.4.2",
     "rollup-plugin-node-resolve": "^3.0.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "ts-node": "^3.3.0",
