+++ src/lib/package.json
@@ -1,11 +1,11 @@
 {
   "name": "ng2-adsense",
-  "version": "4.0.6",
+  "version": "4.0.8",
   "publishConfig": {
     "tag": "next"
   },
   "description": "Adsense for Angular",
-  "main": "./bundles/ngadsense.umd.js",
+  "main": "./bundles/ng2-adsense.umd.js",
   "module": "./index.js",
   "es2015": "./es2015/index.js",
   "typings": "./index.d.ts",
